refactor(edge): split admin-plane auth into admin_identity, admin_auth, and admin_audit
