+++ crates/edge/src/quic_listener/control_api/admin_auth.rs
@@ -1,14 +1,10 @@
-use std::{collections::BTreeSet, net::SocketAddr};
-
 use ::http::{Method, header};
 use bytes::Bytes;
 use http_body_util::Full;
-use serde::Serialize;
-use subtle::ConstantTimeEq;
-use x509_parser::{extensions::GeneralName, parse_x509_certificate, prelude::X509Certificate};
 
 use super::{
-    security::{ControlApiIdentitySourcePolicy, ControlApiSecurityPolicy},
+    admin_identity::{AdminIdentity, AdminRole, ControlApiRequestContext},
+    security::ControlApiSecurityPolicy,
     state::{ControlApiPaths, ControlApiState},
     *,
 };
@@ -58,56 +54,6 @@ impl ControlApiRoute {
     }
 }
 
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Serialize)]
-#[serde(rename_all = "snake_case")]
-pub(super) enum AdminRole {
-    Viewer,
-    Operator,
-    Admin,
-}
-
-impl From<spooky_config::config::ControlApiRole> for AdminRole {
-    fn from(value: spooky_config::config::ControlApiRole) -> Self {
-        match value {
-            spooky_config::config::ControlApiRole::Viewer => Self::Viewer,
-            spooky_config::config::ControlApiRole::Operator => Self::Operator,
-            spooky_config::config::ControlApiRole::Admin => Self::Admin,
-        }
-    }
-}
-
-#[derive(Clone, Debug, Eq, PartialEq, Serialize)]
-#[serde(rename_all = "snake_case")]
-pub(super) enum AdminAuthnMechanism {
-    BearerToken,
-    MutualTls,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq, Serialize)]
-pub(super) struct AdminIdentity {
-    pub(super) actor_id: Option<String>,
-    pub(super) authn_mechanisms: Vec<AdminAuthnMechanism>,
-    pub(super) roles: Vec<AdminRole>,
-    pub(super) peer_addr: Option<SocketAddr>,
-    pub(super) mtls_subject: Option<String>,
-    pub(super) mtls_san: Vec<String>,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub(super) struct ControlApiRequestContext {
-    pub(super) peer_addr: SocketAddr,
-    pub(super) mtls_identity: Option<AdminMtlsIdentity>,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub(super) struct AdminMtlsIdentity {
-    pub(super) subject: Option<String>,
-    pub(super) common_name: Option<String>,
-    pub(super) san_dns: Vec<String>,
-    pub(super) san_uri: Vec<String>,
-    pub(super) roles: Vec<AdminRole>,
-}
-
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub(super) enum AuthorizationDecision {
     Allow {
@@ -129,12 +75,6 @@ enum AuthenticationOutcome {
     Authenticated(AdminIdentity),
     Missing,
     Invalid(&'static str),
-}
-
-#[derive(Clone, Debug)]
-struct AdminTokenMatch {
-    actor_id: Option<String>,
-    role: AdminRole,
 }
 
 impl QUICListener {
@@ -198,99 +138,6 @@ impl QUICListener {
         Some(token)
     }
 
-    pub(super) fn build_control_api_request_context(
-        peer_addr: SocketAddr,
-        peer_certificates: Option<&[CertificateDer<'static>]>,
-        identity_source: Option<&ControlApiIdentitySourcePolicy>,
-    ) -> ControlApiRequestContext {
-        let mtls_identity = peer_certificates
-            .and_then(|certs| certs.first())
-            .and_then(|cert| Self::parse_admin_mtls_identity(cert.as_ref(), identity_source));
-        ControlApiRequestContext {
-            peer_addr,
-            mtls_identity,
-        }
-    }
-
-    fn parse_admin_mtls_identity(
-        cert_der: &[u8],
-        identity_source: Option<&ControlApiIdentitySourcePolicy>,
-    ) -> Option<AdminMtlsIdentity> {
-        let (_, cert) = parse_x509_certificate(cert_der).ok()?;
-        let subject = Some(cert.subject().to_string());
-        let mut common_name = None;
-        for cn in cert.subject().iter_common_name() {
-            if let Ok(value) = cn.as_str() {
-                common_name = Some(value.to_string());
-                break;
-            }
-        }
-
-        let mut san_dns = Vec::new();
-        let mut san_uri = Vec::new();
-        if let Ok(Some(san)) = cert.subject_alternative_name() {
-            for name in &san.value.general_names {
-                match name {
-                    GeneralName::DNSName(dns) => san_dns.push(dns.to_string()),
-                    GeneralName::URI(uri) => san_uri.push(uri.to_string()),
-                    _ => {}
-                }
-            }
-        }
-
-        let roles = identity_source
-            .and_then(|policy| policy.role_attribute.as_deref())
-            .map(|attribute| Self::extract_roles_from_subject_attribute(&cert, attribute))
-            .unwrap_or_default();
-
-        Some(AdminMtlsIdentity {
-            subject,
-            common_name,
-            san_dns,
-            san_uri,
-            roles,
-        })
-    }
-
-    fn extract_roles_from_subject_attribute(
-        cert: &X509Certificate<'_>,
-        attribute: &str,
-    ) -> Vec<AdminRole> {
-        let mut roles = Vec::new();
-        for attr in cert.subject().iter_attributes() {
-            let short_name = attr.attr_type().to_id_string();
-            let matches = short_name.eq_ignore_ascii_case(attribute)
-                || Self::matches_subject_attribute_alias(&short_name, attribute);
-            if !matches {
-                continue;
-            }
-            if let Ok(value) = attr.as_str()
-                && let Some(role) = Self::parse_admin_role(value)
-            {
-                roles.push(role);
-            }
-        }
-        roles
-    }
-
-    fn matches_subject_attribute_alias(actual: &str, requested: &str) -> bool {
-        match requested.to_ascii_lowercase().as_str() {
-            "cn" => actual == "2.5.4.3",
-            "o" => actual == "2.5.4.10",
-            "ou" => actual == "2.5.4.11",
-            _ => false,
-        }
-    }
-
-    fn parse_admin_role(value: &str) -> Option<AdminRole> {
-        match value.trim().to_ascii_lowercase().as_str() {
-            "viewer" => Some(AdminRole::Viewer),
-            "operator" => Some(AdminRole::Operator),
-            "admin" => Some(AdminRole::Admin),
-            _ => None,
-        }
-    }
-
     fn authenticate_control_api_request<B>(
         req: &::http::Request<B>,
         security: &ControlApiSecurityPolicy,
@@ -312,86 +159,14 @@ impl QUICListener {
             None => None,
         };
 
-        let mtls_identity = request_ctx
-            .as_ref()
-            .and_then(|ctx| ctx.mtls_identity.as_ref())
-            .cloned();
-
-        if token_match.is_none() && mtls_identity.is_none() {
-            return AuthenticationOutcome::Missing;
+        match Self::build_admin_identity(
+            request_ctx,
+            token_match,
+            security.identity_source.as_ref(),
+        ) {
+            Some(identity) => AuthenticationOutcome::Authenticated(identity),
+            None => AuthenticationOutcome::Missing,
         }
-
-        let mut roles = BTreeSet::new();
-        let mut mechanisms = Vec::new();
-        let mut actor_id = None;
-        if let Some(token) = token_match {
-            mechanisms.push(AdminAuthnMechanism::BearerToken);
-            roles.insert(token.role);
-            actor_id = token.actor_id.or(actor_id);
-        }
-        if let Some(mtls) = mtls_identity.as_ref() {
-            mechanisms.push(AdminAuthnMechanism::MutualTls);
-            for role in &mtls.roles {
-                roles.insert(*role);
-            }
-            actor_id = actor_id.or_else(|| {
-                Self::actor_id_from_mtls_identity(mtls, security.identity_source.as_ref())
-            });
-        }
-
-        AuthenticationOutcome::Authenticated(AdminIdentity {
-            actor_id,
-            authn_mechanisms: mechanisms,
-            roles: roles.into_iter().collect(),
-            peer_addr: request_ctx.as_ref().map(|ctx| ctx.peer_addr),
-            mtls_subject: mtls_identity.as_ref().and_then(|identity| identity.subject.clone()),
-            mtls_san: mtls_identity
-                .as_ref()
-                .map(|identity| {
-                    identity
-                        .san_dns
-                        .iter()
-                        .cloned()
-                        .chain(identity.san_uri.iter().cloned())
-                        .collect()
-                })
-                .unwrap_or_default(),
-        })
-    }
-
-    fn actor_id_from_mtls_identity(
-        mtls: &AdminMtlsIdentity,
-        identity_source: Option<&ControlApiIdentitySourcePolicy>,
-    ) -> Option<String> {
-        match identity_source
-            .map(|policy| policy.kind.trim().to_ascii_lowercase())
-            .as_deref()
-        {
-            Some("mtls_subject_cn") => mtls.common_name.clone(),
-            Some("mtls_san_dns") => mtls.san_dns.first().cloned(),
-            Some("mtls_san_uri") => mtls.san_uri.first().cloned(),
-            Some("mtls_subject") => mtls.subject.clone(),
-            _ => mtls.common_name.clone().or_else(|| mtls.subject.clone()),
-        }
-    }
-
-    fn control_api_token_match(
-        provided: &str,
-        security: &ControlApiSecurityPolicy,
-    ) -> Option<AdminTokenMatch> {
-        let mut matched = None;
-        let mut matched_any = 0u8;
-        for entry in security.bearer_tokens.iter() {
-            let is_match = bool::from(provided.as_bytes().ct_eq(entry.token.as_bytes())) as u8;
-            if is_match == 1 {
-                matched = Some(AdminTokenMatch {
-                    actor_id: entry.actor_id.clone(),
-                    role: AdminRole::from(entry.role),
-                });
-            }
-            matched_any |= is_match;
-        }
-        (matched_any == 1).then_some(matched).flatten()
     }
 
     #[cfg(test)]
