+++ crates/edge/src/quic_listener/control_api/http.rs
@@ -10,36 +10,36 @@ impl QUICListener {
             Err(response) => return *response,
         };
         match route {
-            super::auth::ControlApiRoute::Health => Self::render_control_api_health(state),
-            super::auth::ControlApiRoute::Ready => Self::render_control_api_ready(state),
-            super::auth::ControlApiRoute::Runtime => {
+            super::admin_auth::ControlApiRoute::Health => Self::render_control_api_health(state),
+            super::admin_auth::ControlApiRoute::Ready => Self::render_control_api_ready(state),
+            super::admin_auth::ControlApiRoute::Runtime => {
                 Self::render_control_api_runtime_snapshot(state)
             }
-            super::auth::ControlApiRoute::RuntimeValidate => {
+            super::admin_auth::ControlApiRoute::RuntimeValidate => {
                 Self::handle_control_api_runtime_validate(req, state).await
             }
-            super::auth::ControlApiRoute::RuntimePreview => {
+            super::admin_auth::ControlApiRoute::RuntimePreview => {
                 Self::handle_control_api_runtime_preview(req, state).await
             }
-            super::auth::ControlApiRoute::RuntimeActivate => {
+            super::admin_auth::ControlApiRoute::RuntimeActivate => {
                 Self::handle_control_api_runtime_activate(req, state).await
             }
-            super::auth::ControlApiRoute::ReloadRuntime => {
+            super::admin_auth::ControlApiRoute::ReloadRuntime => {
                 Self::handle_control_api_runtime_reload(req, state).await
             }
-            super::auth::ControlApiRoute::RuntimeRollback => {
+            super::admin_auth::ControlApiRoute::RuntimeRollback => {
                 Self::handle_control_api_runtime_rollback(req, state).await
             }
-            super::auth::ControlApiRoute::RuntimeHistory => {
+            super::admin_auth::ControlApiRoute::RuntimeHistory => {
                 Self::render_control_api_runtime_history(state)
             }
-            super::auth::ControlApiRoute::RuntimeHistoryGeneration(generation) => {
+            super::admin_auth::ControlApiRoute::RuntimeHistoryGeneration(generation) => {
                 Self::render_control_api_runtime_history_generation(state, generation)
             }
-            super::auth::ControlApiRoute::ReloadCerts => {
+            super::admin_auth::ControlApiRoute::ReloadCerts => {
                 Self::handle_control_api_reload_certs(state)
             }
-            super::auth::ControlApiRoute::Restart => Self::handle_control_api_restart(state),
+            super::admin_auth::ControlApiRoute::Restart => Self::handle_control_api_restart(state),
         }
     }
 }
